# docs/swagger.yaml
definitions:
  request.CreateTodoReq:
    properties:
      name:
        type: string
    type: object
  request.SignInReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  request.SignUpReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  request.UpdateTodoReq:
    properties:
      name:
        type: string
    type: object
info:
  contact: {}
  termsOfService: http://swagger.io/terms/
  title: Product API Gateway
  version: 1.0.4
paths:
  /api/v1/auth/signin:
    post:
      consumes:
      - application/json
      description: SignIn
      parameters:
      - description: SignIn user
        in: body
        name: signin
        required: true
        schema:
          $ref: '#/definitions/request.SignInReq'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: SignIn
      tags:
      - Auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: SignUp
      parameters:
      - description: SignUp user
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/request.SignUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: SignUp
      tags:
      - Auth
  /api/v1/todos:
    get:
      consumes:
      - application/json
      description: Fetch Todo
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - JwtToken: []
      summary: Fetch Todo
      tags:
      - Todos
    post:
      consumes:
      - application/json
      description: Create Todo
      parameters:
      - description: Todo to create
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/request.CreateTodoReq'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - JwtToken: []
      summary: Create Todo
      tags:
      - Todos
  /api/v1/todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Todo
      parameters:
      - description: todo id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - JwtToken: []
      summary: Delete Todo
      tags:
      - Todos
    get:
      consumes:
      - application/json
      description: Get Todo
      parameters:
      - description: todo id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - JwtToken: []
      summary: Get Todo
      tags:
      - Todos
    put:
      consumes:
      - application/json
      description: Update Todo
      parameters:
      - description: todo id
        in: path
        name: id
        required: true
        type: string
      - description: Todo to update
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTodoReq'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - JwtToken: []
      summary: Update Todo
      tags:
      - Todos
securityDefinitions:
  JwtToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"